Check if restarts are directories before deletion
Some users will add a symbolic link to the restart directory of another
run if they are forking an experiment.  shutil.rmtree() fails if the
target is not a directory, and will cause a crash.

We now check to see if the restart directory is a directory before
calling rmtree().

I decided not to use an exception because only directories will be
treated as acceptable for deletion.  Anything else will be treated as a
foreign non-payu file, and will be therefore be preserved.  (All bets
are off for non-payu directories).

+++ payu/experiment.py
@@ -598,7 +598,10 @@ class Experiment(object):
                      res_idx <= (self.counter - restart_history))):
 
                 res_path = os.path.join(self.archive_path, res_dir)
-                shutil.rmtree(res_path)
+
+                # Only delete real directories; ignore symbolic restart links
+                if os.path.isdir(res_path):
+                    shutil.rmtree(res_path)
 
         if self.config.get('collate', True):
             cmd = 'payu collate -i {}'.format(self.counter)
